requirements.txt
Pillow
streamlit
plotly
numpy
requests
json
pathlib
streamlit-lottie